+++ src/libs/geo/src/location.cpp
@@ -35,7 +35,7 @@ geo::Location geo::square(const geo::Location &corner, double diag) {
 }
 
 /* https://stackoverflow.com/questions/7222382/get-lat-long-given-current-point-distance-and-bearing */
-void geo::Location::move(const unsigned long new_time, const double distance, const double bearing) {
+void geo::Location::move(const double new_time, const double distance, const double bearing) {
     auto lat_origin = geo::deg2rad(this->latitude);
     auto lon_origin = geo::deg2rad(this->longitude);
     auto brng = geo::deg2rad(bearing);
@@ -58,19 +58,19 @@ double geo::Location::get_longitude() const {
     return longitude;
 }
 
-unsigned long geo::Location::get_time() const {
+double geo::Location::get_time() const {
     return time;
 }
 
-void geo::Location::set_time(unsigned long new_time) {
+void geo::Location::set_time(double new_time) {
     geo::Location::time = new_time;
 }
 
 geo::Location::Location(double latitude, double longitude) : latitude(latitude), longitude(longitude),
                                                              time(0) {}
 
-geo::Location::Location(unsigned long time, double latitude, double longitude) : time(time), latitude(latitude),
-                                                                                 longitude(longitude) {}
+geo::Location::Location(double time, double latitude, double longitude) : time(time), latitude(latitude),
+                                                                          longitude(longitude) {}
 
 bool geo::Location::operator<(const geo::Location &rhs) const {
     return time < rhs.time;
